Rename test_base_solver.py test_rime_solver.py
